requirements.txt
dukascopy-python
pandas
tqdm
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
